Move library-attributes because it is only used by Dashboard.
